Refactor Next.js config and search page layout
The Next.js config file has been updated to include an experimental flag for `missingSuspenseWithCSRBailout`. This flag is set to `false` to disable the bailout behavior.

In the search page layout, an extra line has been added to import React. This change is necessary for the proper functioning of the layout component.